crew/requirements.txt
crewai>=0.80.0
crewai-tools>=0.30.0
anthropic>=0.40.0
python-dotenv>=1.0.0
pydantic>=2.5.0
